Move to relying on tmpfs to solve game limit issues without writing NAND ref #14 ref #10 ref #9 ref #3
Not currently working. Various iterations of this yield the same thing on my system. The covers show properly for 25 games, but it's almost as if the .pcsx directory for each game is dumped and recreated on start of a game. Dump of FS before game started: https://pastebin.com/raw/kMrh7A3S

This is cleared out once a game is started for the second time, and pcsx.cfg is removed resulting in non-responsive controls. Might just be my console?