Rewrote GameManagerService. Trigger generate folders (WIP)
Rewrote the GameManagerService to better serve folders and to clean up the code. Previously when a game was added, the files were moved right away. This causes problems with database rebuilds where file movements are not required. Since folders are essentially creating many databases, possibly after the files have already been moved, this was required in order to simplify database generation.